# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: 商用地理信息虚假内容检查与反投喂平台 API
  description: |
    ## 概述
    本API提供地理信息虚假内容检测服务，支持单条/批量检测、规则管理、审核标注、统计报表等功能。
    
    ## 认证方式
    - **JWT Token**: 用于Web端用户认证，在Header中添加 `Authorization: Bearer <token>`
    - **API Key**: 用于服务端对接，在Header中添加 `X-API-Key: <api_key>`
    
    ## 请求限制
    - 基础版: 10,000次/天
    - 专业版: 50,000次/天
    - 企业版: 不限制
    
    ## 错误码
    | 错误码 | 说明 |
    |--------|------|
    | 400 | 请求参数错误 |
    | 401 | 未授权/Token过期 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 429 | 请求频率超限 |
    | 500 | 服务器内部错误 |
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@geo-fake-detection.com
servers:
  - url: https://api.geo-fake-detection.com/v1
    description: 生产环境
  - url: https://api-staging.geo-fake-detection.com/v1
    description: 测试环境
...
paths:
...
  /stats/overview:
    get:
      tags: [Stats]
      summary: 总览统计
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, week, month, custom]
            default: today
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsOverview'
  /stats/trend:
    get:
      tags: [Stats]
      summary: 趋势分析
      security:
        - bearerAuth: []
      parameters:
        - name: metric
          in: query
          schema:
            type: string
            enum: [total, fake_rate, risk_distribution, api_calls]
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        time:
                          type: string
                        value:
                          type: number
  /stats/distribution:
    get:
      tags: [Stats]
      summary: 分布统计
      security:
        - bearerAuth: []
      parameters:
        - name: dimension
          in: query
          required: true
          schema:
            type: string
            enum: [risk_level, content_type, geo_source, device]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        count:
                          type: integer
                        percentage:
                          type: number
  /stats/risk:
    get:
      tags: [Stats]
      summary: 风险分析
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAnalysis'

  # ============================================
  # 告警管理
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    StatsOverview:
      type: object
      properties:
        total_records:
          type: integer
        fake_records:
          type: integer
        normal_records:
          type: integer
        fake_rate:
          type: number
        high_risk_count:
          type: integer
        medium_risk_count:
          type: integer
        low_risk_count:
          type: integer
        avg_suspicion_score:
          type: number
        api_calls:
          type: integer
        review_pending:
          type: integer
        review_completed:
          type: integer
    RiskAnalysis:
      type: object
      properties:
        risk_distribution:
          type: object
          properties:
            high:
              type: integer
            medium:
              type: integer
            low:
              type: integer
        top_risk_devices:
          type: array
          items:
            type: object
            properties:
              device_id:
                type: string
              fake_count:
                type: integer
              total_count:
                type: integer
        top_risk_users:
          type: array
          items:
            type: object
            properties:
              user_id:
                type: string
              fake_count:
                type: integer
              total_count:
                type: integer
        anomaly_types:
          type: object
          properties:
            geo_anomaly:
              type: integer
            text_anomaly:
              type: integer
            duplicate:
              type: integer
            cluster_anomaly:
              type: integer
...